chore(adr-012): clippy clean — remove unnecessary cast in dwq_activation
cargo clippy --tests on the worktree surfaces 220 lint errors
total, but only ONE in code authored under ADR-012 P9b:
  src/quantize/dwq_activation.rs:220
    head_dim: Some(hidden_size as u32 / 4)
  hidden_size is already u32, the cast is redundant.

Per "be careful not to clobber" — the other 219 lints pre-date
the P9b worktree and live in files we didn't substantially edit
(gpu_full_attn, gpu_delta_net, models/qwen35/* convert side, etc.).
Out of scope for this commit; left for a dedicated clippy sweep
session.

P9b authored files now clippy-clean:
  ✅ src/backends/gguf.rs (P9b.1 helper + smoke test)
  ✅ src/quantize/dwq_activation.rs (P9b.3a real impl + 7 tests)
  ✅ src/quantize/dwq.rs (P9b.3a refactor: extract internal entry)
  ✅ src/main.rs (P9b.2/3b two-pass branch — diff is small)
  ✅ tests/convert_qwen35_two_pass_capture.rs (P9b.6, new file)
  ✅ tests/convert_qwen35_real_activation_capture.rs (P9b.7 doc-only)
  ✅ Cargo.toml (tempfile promotion)

Tests still green: 1117 passing / 0 failing / 9 ignored.

+++ src/quantize/dwq_activation.rs
@@ -217,7 +217,7 @@ mod tests {
             explicit_layer_types: None,
             full_attention_interval: None,
             attn_output_gate: None,
-            head_dim: Some(hidden_size as u32 / 4),
+            head_dim: Some(hidden_size / 4),
             partial_rotary_factor: None,
             rope_parameters: None,
             linear_conv_kernel_dim: None,
